requirements.txt
pycryptodome
Pillow
